requirements.txt
pycryptodome
Pillow
